naive

+++ src/doradd/config.hpp
@@ -21,7 +21,7 @@ using ts_type = std::chrono::time_point<std::chrono::system_clock>;
 #define L1D_LOCALITY 3
 
 #ifdef LOG_SCHED_OHEAD
-using log_arr_type = std::vector<std::tuple<uint32_t, uint32_t>>;
+using log_arr_type = std::vector<std::tuple<uint64_t, uint64_t>>;
 #else
-using log_arr_type = std::vector<uint32_t>;
+using log_arr_type = std::vector<uint64_t>;
 #endif
